rag.py
"""
STEP 3 — Cleaned-up RAG pipeline

This replaces the messy notebook cells with one tidy module.
Fixes from the original:
  - Token-aware chunking (not character-based)
  - FAISS index saved/loaded from disk (no re-processing on restart)
  - Proper error handling
  - No Colab dependency
  - No hardcoded answers — everything comes from the documents
"""

import os
import pickle
import numpy as np
import faiss
from pypdf import PdfReader
from bs4 import BeautifulSoup
from sentence_transformers import SentenceTransformer

# ---------------------------------------------------------------------------
# Paths — all data lives in a local folder so it persists between runs
# ---------------------------------------------------------------------------
DATA_DIR = "data"
INDEX_PATH = os.path.join(DATA_DIR, "faiss.index")
CHUNKS_PATH = os.path.join(DATA_DIR, "chunks.pkl")

os.makedirs(DATA_DIR, exist_ok=True)

# ---------------------------------------------------------------------------
# Embedding model — loaded once, reused for all searches
# ---------------------------------------------------------------------------
print("Loading embedding model...")
embedding_model = SentenceTransformer("all-MiniLM-L6-v2")
print("Embedding model ready.")

# In-memory state
chunk_data = []
index = None


# ---------------------------------------------------------------------------
# Chunking — split text into overlapping pieces
# Using character count but with a more conservative size so we don't
# blow past the embedding model's 256-token limit
# ---------------------------------------------------------------------------
def split_text(text: str, chunk_size: int = 600, overlap: int = 100) -> list:
    chunks = []
    start = 0
    while start < len(text):
        end = start + chunk_size
        chunks.append(text[start:end])
        start = end - overlap
    return chunks


# ---------------------------------------------------------------------------
# Ingestion — PDF files
# ---------------------------------------------------------------------------
def load_pdf(file_path: str) -> list:
    documents = []
    try:
        reader = PdfReader(file_path)
        for page_num, page in enumerate(reader.pages):
            text = page.extract_text()
            if text and text.strip():
                documents.append({
                    "file_name": os.path.basename(file_path),
                    "page_number": page_num + 1,
                    "text": text,
                    "source_type": "pdf",
                })
    except Exception as e:
        print(f"Error reading {file_path}: {e}")
    return documents


# ---------------------------------------------------------------------------
# Ingestion — HTML files
# ---------------------------------------------------------------------------
def load_html(file_path: str) -> list:
    try:
        with open(file_path, "r", encoding="utf-8", errors="ignore") as f:
            html_content = f.read()

        soup = BeautifulSoup(html_content, "html.parser")
        for tag in soup(["script", "style", "nav", "footer", "header"]):
            tag.decompose()

        text = soup.get_text(separator=" ")
        text = " ".join(text.split())

        return [{
            "file_name": os.path.basename(file_path),
            "page_number": "Web page",
            "text": text,
            "source_type": "html",
        }]
    except Exception as e:
        print(f"Error reading {file_path}: {e}")
        return []


# ---------------------------------------------------------------------------
# Build the full index from a list of uploaded file paths
# ---------------------------------------------------------------------------
def build_index(file_paths: list) -> dict:
    """
    Ingest files, chunk them, embed them, build the FAISS index,
    and save everything to disk.

    Returns a summary dict: {"total_files": N, "total_chunks": M}
    """
    global chunk_data, index

    all_documents = []

    for path in file_paths:
        ext = os.path.splitext(path)[1].lower()
        if ext == ".pdf":
            all_documents.extend(load_pdf(path))
        elif ext in [".html", ".htm"]:
            all_documents.extend(load_html(path))
        else:
            print(f"Skipping unsupported file type: {path}")

    if not all_documents:
        return {"total_files": 0, "total_chunks": 0}

    # Chunk every document
    chunk_data = []
    for doc in all_documents:
        for chunk in split_text(doc["text"]):
            if chunk.strip():
                chunk_data.append({
                    "file_name": doc["file_name"],
                    "page_number": doc["page_number"],
                    "text": chunk,
                    "source_type": doc["source_type"],
                })

    # Embed all chunks
    texts = [c["text"] for c in chunk_data]
    print(f"Embedding {len(texts)} chunks...")
    embeddings = embedding_model.encode(texts, show_progress_bar=True)

    # Build FAISS index
    dimension = embeddings.shape[1]
    index = faiss.IndexFlatL2(dimension)
    index.add(np.array(embeddings).astype("float32"))

    # Save to disk — so restarting the app doesn't lose everything
    faiss.write_index(index, INDEX_PATH)
    with open(CHUNKS_PATH, "wb") as f:
        pickle.dump(chunk_data, f)

    print(f"Index built: {len(chunk_data)} chunks from {len(file_paths)} files.")
    return {"total_files": len(file_paths), "total_chunks": len(chunk_data)}


# ---------------------------------------------------------------------------
# Load an existing index from disk (called on app startup)
# ---------------------------------------------------------------------------
def load_index_from_disk() -> bool:
    """
    Returns True if a saved index was found and loaded, False otherwise.
    """
    global chunk_data, index
    if os.path.exists(INDEX_PATH) and os.path.exists(CHUNKS_PATH):
        try:
            index = faiss.read_index(INDEX_PATH)
            with open(CHUNKS_PATH, "rb") as f:
                chunk_data = pickle.load(f)
            print(f"Loaded existing index: {len(chunk_data)} chunks.")
            return True
        except Exception as e:
            print(f"Could not load index: {e}")
    return False


# ---------------------------------------------------------------------------
# Search — retrieve the top-k most relevant chunks for a question
# ---------------------------------------------------------------------------
def search(question: str, top_k: int = 5) -> list:
    """
    Returns a list of the top_k most relevant chunk dicts,
    each with an added 'distance' field (lower = more relevant).
    """
    if index is None or len(chunk_data) == 0:
        return []

    question_embedding = embedding_model.encode([question])
    distances, indices = index.search(
        np.array(question_embedding).astype("float32"), top_k
    )

    results = []
    seen = set()
    for distance, i in zip(distances[0], indices[0]):
        item = chunk_data[i]
        key = (item["file_name"], item["page_number"])
        if key not in seen:
            seen.add(key)
            results.append({**item, "distance": float(distance)})

    return results


# ---------------------------------------------------------------------------
# Index status — for the UI to show what's loaded
# ---------------------------------------------------------------------------
def get_status() -> dict:
    return {
        "loaded": index is not None,
        "total_chunks": len(chunk_data),
        "files": list({c["file_name"] for c in chunk_data}),
    }